requirements.txt
automationhat==0.1.0
backports-abc==0.5
blinker==1.3
blinkt==0.1.2
buttonshim==0.0.2
cachetools==1.1.6
Cap1xxx==0.1.3
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.3.7
colorzero==1.1
configobj==5.0.6
cryptography==1.7.1
drumhat==0.1.0
enum34==1.1.6
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==1.0.3
fourletterphat==0.1.0
futures==3.0.5
gpiozero==1.5.0
idna==2.8
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10.1
keyring==10.1
keyrings.alt==1.3
MarkupSafe==1.1.1
microdotphat==0.2.1
Mopidy==2.2.2
Mopidy-ALSAMixer==1.0.3
Mopidy-InternetArchive==2.0.3
Mopidy-Iris==3.37.1
Mopidy-Local-Images==1.0.0
Mopidy-Local-SQLite==1.0.0
Mopidy-Podcast==2.0.3
Mopidy-Podcast-iTunes==2.0.0
Mopidy-SomaFM==1.1.0
Mopidy-SoundCloud==2.1.0
Mopidy-Youtube==2.0.0
mote==0.0.4
motephat==0.0.2
numpy==1.12.1
oauthlib==2.0.1
pantilthat==0.0.7
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
picraft==1.0
piglow==1.2.4
pigpio==1.38
Pillow==4.0.0
pyalsaaudio==0.7
pyasn1==0.1.9
pycrypto==2.6.1
pygame==1.9.3
pygobject==3.22.0
pyinotify==0.9.6
PyJWT==1.4.2
Pykka==2.0.0
pylast==1.6.0
pyOpenSSL==16.2.0
pyserial==3.2.1
pyxdg==0.25
rainbowhat==0.1.0
requests==2.22.0
requests-oauthlib==0.7.0
RPi.GPIO==0.6.5
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
sense-hat==2.2.0
simplejson==3.10.0
singledispatch==3.4.0.3
six==1.12.0
skywriter==0.0.7
sn3218==1.2.7
spidev==3.3
tornado==4.5.3
touchphat==0.0.1
twython==3.4.0
unicornhathd==0.0.4
uritools==2.2.0
urllib3==1.25.2
Werkzeug==0.15.4
